# Add primitive methods to TDigest stats