--- a/stats/src/test/java/io/airlift/stats/TestTDigest.java
+++ b/stats/src/test/java/io/airlift/stats/TestTDigest.java
@@ -16,6 +16,7 @@ package io.airlift.stats;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Lists;
 import com.google.common.io.Resources;
+import com.google.common.primitives.Doubles;
 import io.airlift.slice.Slices;
 import org.testng.annotations.Test;
 
@@ -41,7 +42,10 @@ public class TestTDigest
         assertTrue(Double.isNaN(digest.valueAt(0.5)));
         assertTrue(Double.isNaN(digest.getMin()));
         assertTrue(Double.isNaN(digest.getMax()));
-        assertEquals(digest.valuesAt(ImmutableList.of(0.1, 0.2, 0.5, 0.9)), ImmutableList.of(Double.NaN, Double.NaN, Double.NaN, Double.NaN));
+        double[] quantiles = new double[]{0.1, 0.2, 0.5, 0.9};
+        double[] expected = new double[]{Double.NaN, Double.NaN, Double.NaN, Double.NaN};
+        assertEquals(digest.valuesAt(quantiles), expected);
+        assertEquals(digest.valuesAt(Doubles.asList(quantiles)), Doubles.asList(expected));
     }
 
     @Test
@@ -73,7 +77,7 @@ public class TestTDigest
         assertEquals(digest.valueAt(0.9), 19.0);
         assertEquals(digest.valueAt(0.949999999), 19.0);
         assertEquals(digest.valueAt(0.95), 1_000_000.0);
-        assertEquals(digest.valuesAt(ImmutableList.of(0.89999999, 0.9, 0.949999999, 0.95)), ImmutableList.of(18.0, 19.0, 19.0, 1_000_000.0));
+        assertEquals(digest.valuesAt(0.89999999, 0.9, 0.949999999, 0.95), new double[]{18.0, 19.0, 19.0, 1_000_000.0});
     }
 
     @Test
@@ -87,7 +91,7 @@ public class TestTDigest
 
         assertEquals(digest.valueAt(0.998), 999.0);
         assertEquals(digest.valueAt(0.999), 1_000_000.0);
-        assertEquals(digest.valuesAt(ImmutableList.of(0.998, 0.999)), ImmutableList.of(999.0, 1_000_000.0));
+        assertEquals(digest.valuesAt(0.998, 0.999), new double[]{999.0, 1_000_000.0});
     }
 
     @Test
@@ -102,10 +106,10 @@ public class TestTDigest
         assertEquals(digest.valueAt(0.5 - 1e-10), 20, 1e-10);
         assertEquals(digest.valueAt(0.5), 32, 1e-10);
 
-        List<Double> quantiles = ImmutableList.of(0.25 - 1e-10, 0.25, 0.4, 0.5 - 1e-10, 0.5);
-        List<Double> values = digest.valuesAt(quantiles);
-        for (int i = 0; i < quantiles.size(); i++) {
-            assertEquals(values.get(i), digest.valueAt(quantiles.get(i)));
+        double[] quantiles = new double[]{0.25 - 1e-10, 0.25, 0.4, 0.5 - 1e-10, 0.5};
+        double[] values = digest.valuesAt(quantiles);
+        for (int i = 0; i < quantiles.length; i++) {
+            assertEquals(values[i], digest.valueAt(quantiles[i]));
         }
     }
 
@@ -144,9 +148,7 @@ public class TestTDigest
         assertEquals(digest.valueAt(0), value, 0.001f);
         assertEquals(digest.valueAt(0.5), value, 0.001f);
         assertEquals(digest.valueAt(1), value, 0.001f);
-        assertEquals(
-                digest.valuesAt(ImmutableList.of(0d, 0.5d, 1d)),
-                ImmutableList.of(digest.valueAt(0), digest.valueAt(0.5), digest.valueAt(1)));
+        assertEquals(digest.valuesAt(0d, 0.5d, 1d), new double[]{digest.valueAt(0), digest.valueAt(0.5), digest.valueAt(1)});
     }
 
     @Test
@@ -160,7 +162,7 @@ public class TestTDigest
         assertEquals(digest.valueAt(0.3), 1.0);
         assertEquals(digest.valueAt(0.9), 2.0);
         assertEquals(digest.valueAt(1), 2.0);
-        assertEquals(digest.valuesAt(ImmutableList.of(0d, 0.3d, 0.9d, 1d)), ImmutableList.of(1.0, 1.0, 2.0, 2.0));
+        assertEquals(digest.valuesAt(0d, 0.3d, 0.9d, 1d), new double[]{1.0, 1.0, 2.0, 2.0});
     }
 
     @Test
@@ -172,7 +174,7 @@ public class TestTDigest
         digest.add(3);
         digest.add(4);
 
-        assertEquals(digest.valuesAt(ImmutableList.of(0d, 0.6d, 1d)), ImmutableList.of(1.0, 3.0, 4.0));
+        assertEquals(digest.valuesAt(0d, 0.6d, 1d), new double[]{1.0, 3.0, 4.0});
     }
 
     @Test
@@ -337,6 +339,8 @@ public class TestTDigest
 
         // empty quantiles list
         assertEquals(ImmutableList.of(), digest.valuesAt(ImmutableList.of()));
+        assertEquals(new double[0], digest.valuesAt());
+        assertEquals(new double[0], digest.valuesAt(new double[0]));
 
         // quantiles not sorted
         assertThrows(IllegalArgumentException.class, () -> digest.valuesAt(ImmutableList.of(0.9, 0.1)));