Adds primitive method `TDigest#valuesAt(double... percentiles)` to avoid unnecessary boxing, and updates usage sites to prefer calling the method without boxing. Also reduces the scope and duration of synchronization during `TimeDistribution` and `Distribution` snapshot creation and redundant work in percentile generation.