@pettyjamesm 2022-10-27
Benchmark results:

**TDigest#valueAt**
```
**Before**
Benchmark                                                           Mode  Cnt     Score      Error   Units
BenchmarkTDigest.benchmarkValueAt                                   avgt   10     7.388 ±    0.117   us/op
BenchmarkTDigest.benchmarkValueAt:·gc.alloc.rate.norm               avgt   10  3552.645 ±    0.331    B/o

**After**
Benchmark                                                           Mode  Cnt    Score      Error   Units
BenchmarkTDigest.benchmarkValueAt                                   avgt   10    6.395 ±    0.047   us/op
BenchmarkTDigest.benchmarkValueAt:·gc.alloc.rate.norm               avgt   10  288.053 ±    0.169    B/op
```

**TDigest#valuesAt**
```
**Before**
Benchmark                                                    (quantileCount)  Mode  Cnt     Score     Error   Units
BenchmarkTDigest.benchmarkValuesAtList                                     1  avgt   10     1.032 ±   0.010   us/op
BenchmarkTDigest.benchmarkValuesAtList:·gc.alloc.rate.norm                 1  avgt   10   400.081 ±   0.045    B/op
BenchmarkTDigest.benchmarkValuesAtList                                     5  avgt   10     1.446 ±   0.014   us/op
BenchmarkTDigest.benchmarkValuesAtList:·gc.alloc.rate.norm                 5  avgt   10   856.149 ±   0.041    B/op
BenchmarkTDigest.benchmarkValuesAtList                                   100  avgt   10     4.002 ±   0.039   us/op
BenchmarkTDigest.benchmarkValuesAtList:·gc.alloc.rate.norm               100  avgt   10  9513.262 ±   0.116    B/op

**After**
Benchmark                                                     (quantileCount)  Mode  Cnt    Score     Error   Units
BenchmarkTDigest.benchmarkValuesAtArray                                     1  avgt   10    0.951 ±   0.021   us/op
BenchmarkTDigest.benchmarkValuesAtArray:·gc.alloc.rate.norm                 1  avgt   10   24.008 ±   0.025    B/op
BenchmarkTDigest.benchmarkValuesAtArray                                     5  avgt   10    1.058 ±   0.007   us/op
BenchmarkTDigest.benchmarkValuesAtArray:·gc.alloc.rate.norm                 5  avgt   10   56.014 ±   0.035    B/op
BenchmarkTDigest.benchmarkValuesAtArray                                   100  avgt   10    1.546 ±   0.018   us/op
BenchmarkTDigest.benchmarkValuesAtArray:·gc.alloc.rate.norm               100  avgt   10  816.143 ±   0.053    B/op
```